requirements.txt
yt-dlp==2023.12.30
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.0
